+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/classification/SoftMaxRegressionTest.java
@@ -64,7 +64,6 @@ public class SoftMaxRegressionTest extends AbstractTest {
         
         SoftMaxRegression.TrainingParameters param = new SoftMaxRegression.TrainingParameters();
         param.setTotalIterations(2000);
-        param.setL1(0.001);
         param.setL2(0.001);
         
         instance.fit(trainingData, param);
@@ -125,8 +124,8 @@ public class SoftMaxRegressionTest extends AbstractTest {
         
         SoftMaxRegression.TrainingParameters param = new SoftMaxRegression.TrainingParameters();
         param.setTotalIterations(30);
-        param.setL1(0.00000001);
-        param.setL2(0.001);
+        param.setL1(0.0001);
+        param.setL2(0.0001);
         
         SoftMaxRegression.ValidationMetrics vm = instance.kFoldCrossValidation(trainingData, param, k);